add util finsih init stamp script

+++ script/init-stamp.js
@@ -5,29 +5,18 @@
 const FS = require("fs");
 const OS = require("os");
 const Path = require("path");
-const Readline = require('readline');
 
 const { exec } = require("child_process");
+const { readlineQuestion } = require("./util");
 
 const blankImagePath = Path.join(__dirname, '..', 'stamp', '$', 'blank', 'blank.png');
 const blankYMLPath = Path.join(__dirname, '..', '_data', 'stamp', 'blank.yml');
 
-const readline = Readline.createInterface(
-    process.stdin,
-    process.stdout,
-);
+(async () => {
 
-readline.question('[~~~~] Stamp Name: ', (stampName) => {
+    const stampName = await readlineQuestion("Stamp Name");
 
     const targetPathFolder = Path.join(__dirname, '..', 'stamp', '$', stampName);
-
-    if (!FS.existsSync(targetPathFolder)) {
-        FS.mkdirSync(targetPathFolder);
-    }
-
-    FS.copyFileSync(blankImagePath, Path.join(targetPathFolder, 'en-US.png'));
-
-    console.log(`[INFO] Stamp Icon ${stampName} created`);
 
     const targetYMLPath = Path.join(__dirname, '..', '_data', 'stamp', `${stampName}.yml`);
 
@@ -36,12 +25,118 @@ readline.question('[~~~~] Stamp Name: ', (stampName) => {
     const currentMonth = (currentDate.getMonth() + 1).toString().padStart(2, '0');
     const currentDay = currentDate.getDate().toString().padStart(2, '0');
 
+    const availableRarity = [
+        "well-known",
+        "uncommon",
+        "worthy",
+        "rare",
+        "dazzle",
+        "sparkling",
+        "priceless",
+        "special",
+    ];
+
+    let chooseRarityIndex = null;
+
+    while (typeof availableRarity[parseInt(chooseRarityIndex)] === 'undefined') {
+
+        console.log("[----] Pick a Rarity:");
+        availableRarity.forEach((method, index) => {
+            console.log(`[----] ${index}: ${method}`);
+        });
+
+        chooseRarityIndex = await readlineQuestion("Rarity (Index)");
+
+        if (typeof availableRarity[parseInt(chooseRarityIndex)] === 'undefined') {
+            console.log("[ERRR] Invalid Index");
+        }
+    }
+
+    console.log(`[INFO] Method: ${availableRarity[parseInt(chooseRarityIndex)]}`);
+
+    const availableMethods = [
+        "regular",
+        "auto",
+        "email",
+        "pull-request",
+        "special",
+    ];
+
+    const pickedMethods = [];
+    console.log("[----] Pick Methods:");
+
+    while (true) {
+
+        console.log(`[----] Method #${pickedMethods.length + 1}:`);
+
+        availableMethods.forEach((method, index) => {
+
+            if (pickedMethods.includes(method)) {
+                return;
+            }
+            console.log(`[----] ${index}: ${method}`);
+        });
+
+        const chooseMethodIndex = await readlineQuestion(`Method #${pickedMethods.length + 1} (Index)`);
+
+        if (chooseMethodIndex === "") {
+            break;
+        }
+
+        const chooseMethodValue = availableMethods[chooseMethodIndex];
+
+        if (!pickedMethods.includes(chooseMethodValue)) {
+
+            pickedMethods.push(chooseMethodValue);
+        } else {
+
+            console.log("[WARN] Already picked, Skipped");
+        }
+
+        console.log(`[INFO] Methods: ${pickedMethods.join(', ')}`);
+    }
+
     const blankYMLData = FS.readFileSync(blankYMLPath, 'utf8');
     const fixedBlankYML = blankYMLData
         .replace('identifier: blank', `identifier: ${stampName}`)
+        .replace('rarity: special', `rarity: ${availableRarity[parseInt(chooseRarityIndex)]}`)
+        .replace('  - special', pickedMethods.map((method) => {
+
+            return `  - ${method}`;
+        }).join('\n'))
         .replace('2022-01-27', `${currentYear}-${currentMonth}-${currentDay}`)
         .replace('https://stamp.sudo.tv/stamp/$/blank/blank.png', `https://stamp.sudo.tv/stamp/$/${stampName}/en-US.png`)
-        .replace('https://stamp.sudo.tv/stamp/$/blank/blank.png', `https://stamp.sudo.tv/stamp/$/${stampName}/en-US.png`);
+        .replace('https://stamp.sudo.tv/stamp/$/blank/blank.png', `https://stamp.sudo.tv/stamp/$/${stampName}/en-US.png`)
+        .replace('label: "空白贴纸"', 'label: "PLACEHOLDER"')
+        .replace('description: "完全空白的贴纸"', 'description: |-\n      PLACEHOLDER')
+        .replace('how-to-get: "哈，这就不知道了"', 'how-to-get: |-\n      PLACEHOLDER')
+        .replace('label: "Blank Stamp"', 'label: "PLACEHOLDER"')
+        .replace('description: "Completely blank stamp"', 'description: |-\n      PLACEHOLDER')
+        .replace('how-to-get: "Ha, I don\'t know then"', 'how-to-get: |-\n      PLACEHOLDER');
+
+    const preview = await readlineQuestion("Preview? (y/n)");
+
+    if (preview === 'y') {
+
+        console.log("[INFO] Preview:");
+        console.log(fixedBlankYML);
+    }
+
+    const confirm = await readlineQuestion("Confirm? (y/n)");
+
+    if (confirm !== 'y') {
+
+        console.log("[INFO] Canceled");
+        return;
+    }
+
+    if (!FS.existsSync(targetPathFolder)) {
+        FS.mkdirSync(targetPathFolder);
+    }
+
+    FS.copyFileSync(blankImagePath, Path.join(targetPathFolder, 'en-US.png'));
+
+    console.log(`[INFO] Stamp Icon ${stampName} created`);
 
     FS.writeFileSync(targetYMLPath, fixedBlankYML);
 
@@ -55,7 +150,10 @@ readline.question('[~~~~] Stamp Name: ', (stampName) => {
         exec(`open ${targetPathFolder}`);
     }
 
+    console.log(`[INFO] Opening ${targetPathFolder}`);
+
     exec(`code ${targetYMLPath}`);
 
-    readline.close();
-});
+    console.log(`[INFO] Opening ${targetPathFolder}/${stampName}.yml`);
+    console.log("[INFO] Finished");
+})();
